requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
pyparsing==2.4.7
PySDL2==0.9.7
pysdl2-dll==2.0.12
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
